+++ PCE_Web/PCE_Web/Views/PasswordChange/ChangePassword.cshtml
@@ -1,4 +1,5 @@
-﻿@model PCE_Web.Controllers.PasswordChangeController.InputModel
+﻿@*@model PCE_Web.Controllers.PasswordChangeController.InputModel*@
+@model InputModel
 @{
     ViewData["Title"] = "Slaptažodžio keitimas";
 }
@@ -32,14 +33,14 @@
             <hr>
 
             <div class="form-group">
-                <label asp-for="Email"><b>El. paštas</b></label>
-                <input asp-for="Email" class="form-control" type="email" placeholder="Įveskite el. paštą"/>
-                <span asp-validation-for="Email" class="text-danger"></span>
+                <label asp-for="EmailModel.Email"><b>El. paštas</b></label>
+                <input asp-for="EmailModel.Email" class="form-control" type="email" placeholder="Įveskite el. paštą"/>
+                <span asp-validation-for="EmailModel.Email" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Code"><b>Kodas</b></label>
-                <input disabled="disabled" asp-for="Code" class="form-control" type="email" placeholder="Įveskite kodą" />
-                <span asp-validation-for="Code" class="text-danger"></span>
+                <label asp-for="CodeModel.Code"><b>Kodas</b></label>
+                <input disabled="disabled" asp-for="CodeModel.Code" class="form-control" type="email" placeholder="Įveskite kodą" />
+                <span asp-validation-for="CodeModel.Code" class="text-danger"></span>
                 @if (ViewBag.ShowMessage == true)
                 {
                     <script type="text/javascript">
@@ -51,14 +52,14 @@
                 }
             </div>
             <div class="form-group">
-                <label asp-for="Password"><b>Slaptažodis</b></label>
-                <input disabled="disabled" asp-for="Password" class="form-control" type="password" placeholder="Įveskite slaptažodį" />
-                <span asp-validation-for="Password" class="text-danger"></span>
+                <label asp-for="PasswordModel.Password"><b>Slaptažodis</b></label>
+                <input disabled="disabled" asp-for="PasswordModel.Password" class="form-control" type="password" placeholder="Įveskite slaptažodį" />
+                <span asp-validation-for="PasswordModel.Password" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="ConfirmPassword"><b>Pakartokite slaptažodį</b></label>
-                <input disabled="disabled" asp-for="ConfirmPassword" class="form-control" type="password" placeholder="Pakartokite slaptažodį" />
-                <span asp-validation-for="ConfirmPassword" class="text-danger"></span>
+                <label asp-for="PasswordModel.ConfirmPassword"><b>Pakartokite slaptažodį</b></label>
+                <input disabled="disabled" asp-for="PasswordModel.ConfirmPassword" class="form-control" type="password" placeholder="Pakartokite slaptažodį" />
+                <span asp-validation-for="PasswordModel.ConfirmPassword" class="text-danger"></span>
             </div>
 
             <button name="submitbtn" type="submit" class="registerbtn">Siųsti</button>
